requirements.txt
attrs==22.1.0
autopep8==1.7.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
cloudevents==1.6.1
deprecation==2.1.0
five9==0.0.3b192
Flask==2.2.2
functions-framework==3.2.0
google-api-core==2.10.1
google-auth==2.11.1
google-cloud-core==2.3.2
google-cloud-firestore==2.7.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
handler-cf-v1==1.0.29
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
platformdirs==2.5.2
properties==0.6.1
proto-plus==1.22.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
toml==0.10.2
urllib3==1.26.12
watchdog==2.1.9
websocket-client==1.4.1
Werkzeug==2.2.2
zeep==4.1.0
